added rabbitmq listener and sender for testing

+++ backend/springbootMS/products/src/main/java/com/spencer/config/InventoryUpdate.java
@@ -6,5 +6,5 @@ import lombok.Data;
 // Message template 3
 @Data
 public class InventoryUpdate {
-	HashMap<Long, Integer> prods;
+	HashMap<Long, Long> prods;
 }
